test: split models.test.ts into catalog.test.ts + discovery.test.ts
- tests/unit/catalog.test.ts: 10 tests for static model data +
  thinking maps (imports catalog.js + thinking.js directly)
- tests/unit/discovery.test.ts: 10 tests for fetchRemoteModels +
  resolveModels (imports discovery.js + catalog.js)
- tests/unit/models.test.ts: reduced to 4 barrel integration tests

Fulfills architecture spec: 'test file splits along the seam.'
101 tests across 11 files.
